Xext: move sleepuntil.c to dix
It's a generic function operating on clients, not extensions specific.
