+++ Xext/xtest/xtest.c
@@ -41,6 +41,7 @@
 #include "dix/inpututils_priv.h"
 #include "dix/request_priv.h"
 #include "dix/screensaver_priv.h"
+#include "dix/sleepuntil.h"
 #include "dix/window_priv.h"
 #include "mi/mi_priv.h"
 #include "mi/mipointer_priv.h"
@@ -57,7 +58,6 @@
 #include "windowstr.h"
 #include "inputstr.h"
 #include "scrnintstr.h"
-#include "sleepuntil.h"
 #include "xkbsrv.h"
 #include "xkbstr.h"
 #include "Xext/xinput/exglobals.h"
